fix(autocomplete): fix issue when asynchrone data
We did a lot of investigation on this issue with @julienmege and didn't find the root cause of it.
We suspect an internal issue of ui-bootstrap typeahead which is not resolved in last tryed version 2.5.0.

Anyway adding some debounce (`typeahead-wait-ms`) fixes the issue. This is actually a double win since it also
enhance the user experience by making less api calls in async use and less processing in sync use.

Closes [BS-16696](https://bonitasoft.atlassian.net/browse/BS-16696)