chore(ci): Push tag to public repository (#39)
Used action to release is not maintained anymore. So I switch for another release-action
To push the tag on the public repository, we use a remote that targets the public repository.

As we use a self-hosted runner, we don't get a fresh install for each run. To avoid Github error Remote already exist when we adding a remote, we need to use this command line:

git config remote.remoteName.url >&- || git remote add remoteName https://github.com/owner/repo.git